extract medication reminders into separate service object
