+++ app/services/experiment_control_service.rb
@@ -61,35 +61,6 @@ class ExperimentControlService
       experiment.running_state!
     end
 
-    # This is not a scientific experiment like active and stale patient experiments. It does not need or have human study
-    # approval and we are not excluding patients under 18 years old. We are leveraging the experimentation framework to
-    # request patients to visit clinics to pick up enough blood pressure medication to last the current covid spike.
-    # We're using the experimentation framework for two reasons:
-    # 1) we want most of the same functionality, including batched daily notification scheduling and the ability to track
-    #    results by leveraging the experimentation and notification models
-    # 2) this is an excellent opportunity to real-world test the experimentation framework
-    def start_medication_reminder_experiment(name, patients_per_day: PATIENTS_PER_DAY)
-      return unless Flipper.enabled?(:live_experiment)
-
-      experiment = Experimentation::Experiment.find_by!(name: name, experiment_type: "medication_reminder")
-      experiment.selecting_state!
-
-      eligible_ids = medication_reminder_patients(experiment)
-      if eligible_ids.any?
-        eligible_ids.shuffle!
-
-        daily_ids = eligible_ids.pop(patients_per_day)
-        daily_patients = Patient.where(id: daily_ids)
-        daily_patients.each do |patient|
-          group = experiment.random_treatment_group
-          schedule_reminders(patient, nil, group, Date.current)
-          group.patients << patient
-        end
-      end
-
-      experiment.running_state!
-    end
-
     protected
 
     def current_patient_candidates(start_date, end_date)
@@ -97,21 +68,6 @@ class ExperimentControlService
         .joins(:appointments)
         .merge(Appointment.status_scheduled)
         .where("appointments.scheduled_date BETWEEN ? and ?", start_date, end_date)
-        .distinct
-        .pluck(:id)
-    end
-
-    def medication_reminder_patients(experiment)
-      Patient
-        .with_hypertension
-        .contactable
-        .includes(treatment_group_memberships: [treatment_group: [:experiment]])
-        .where("experiments.id IS NULL OR experiments.id != ?", experiment.id)
-        .references(:experiment)
-        .where(
-          "NOT EXISTS (SELECT 1 FROM blood_pressures WHERE blood_pressures.patient_id = patients.id AND blood_pressures.device_created_at > ?)",
-          30.days.ago.beginning_of_day
-        )
         .distinct
         .pluck(:id)
     end
