+++ spec/jobs/appointment_notification/worker_spec.rb
@@ -25,9 +25,9 @@ RSpec.describe AppointmentNotification::Worker, type: :job do
       allow(twilio_client).to receive_message_chain("messages.create")
     end
 
-    it "logs but creates nothing when notifications and live_experiment flags are disabled" do
+    it "logs but creates nothing when notifications and experiment flags are disabled" do
       Flipper.disable(:notifications)
-      Flipper.disable(:live_experiment)
+      Flipper.disable(:experiment)
 
       expect(Statsd.instance).to receive(:increment).with("appointment_notification.worker.skipped.feature_disabled")
       expect {
@@ -45,9 +45,9 @@ RSpec.describe AppointmentNotification::Worker, type: :job do
       }.to change { Communication.count }.by(1)
     end
 
-    it "creates communications when live_experiment is enabled" do
+    it "creates communications when experiment is enabled" do
       Flipper.disable(:notifications)
-      Flipper.enable(:live_experiment)
+      Flipper.enable(:experiment)
 
       mock_successful_delivery
